+++ core-java-sample/src/main/java/org/spine3/sample/procman/package-info.java
@@ -19,6 +19,6 @@
  */
 
 @ParametersAreNonnullByDefault
-package org.spine3.sample.processmanager;
+package org.spine3.sample.procman;
 
 import javax.annotation.ParametersAreNonnullByDefault;
