+++ core-java/src/main/java/org/spine3/server/procman/ProcessManager.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.process;
+package org.spine3.server.procman;
 
 import com.google.common.base.Function;
 import com.google.common.collect.ImmutableSet;
